Fix up type checking
The TypeNamespace typeclass now exports declareNamedType, which
handle the common workflow of `type a <...>`.

Tests were added for this type class.

The logic around the pure handling of types was fixed, with tests added.

In particular `type a int` was allowd to both be assigned and assign to
an int, which deviates from Go's behavior. `byte` and `rune` however are
allowed to assign to and from `uint8` and `int32` respectively, because
they are aliases instead of named types.